examples done
